Fixed issue with server not starting, needed to use threads. I'll look into this more later and try and implement it correctly, for now this works.
Added a new endpoint/function for searching on the Spotify API, it's pretty
bare-bones at the moment, will flesh it out later.

Added a label/button and list to the gui to show results, may need to
add widgets for showing different parts of the app(playlists, artist
etc...) Will need to look at the pyQt5 docs on how to do it.