Finally got the player fully working, grabbed 3GP links instead of WEBM
This also solved my issue with grabbing the duration of the files, I've
now implemented the volume bar, progress bar and they are connected to
the player. I still need to setup scrubbing.

Ive added threads to the gui using QThread, this stopped the gui from
freezing when calling my functions outside of the gui. Using the
slots/signals i can send my data back and forth. This seems to be the
best way.

I'll continue to use threading for the server as it's outside the gui
and i only ever need it for the login process.

I've setup right-clicking functionality on the results page, this allows
the user to go to other pages and also play any results(tracks). Still
loads to do here.